Many changes
Added a new abstract ephemeris class. The JPL ephemeris is now an instance of this. Also added a SPICE ephemeris example.
Added a new module for analytical ephemerides (currently, has a Moon ephemeris).
Added a new reference frame transformation module.
Various other updates.

+++ src/crtbp_module.f90
@@ -286,7 +286,7 @@
     real(wp),dimension(6),intent(out) :: dx   !! normalized state derivative \([\dot{\mathbf{r}},\dot{\mathbf{v}}]\)
 
     !local variables:
-    real,dimension(3) :: r1,r2,rb1,rb2,r,v,g
+    real(wp),dimension(3) :: r1,r2,rb1,rb2,r,v,g
     real(wp) :: r13,r23,omm,c1,c2
 
     !extract variables from x vector:
